requirements.txt
# Production dependencies — installed by Streamlit Cloud.
# Keep this list lean. Training-only packages live in requirements-train.txt.

# Web app framework
streamlit==1.50.0

# Google Cloud Storage (video served from public URL; GCS client kept for future use)
google-cloud-storage==3.4.0
google-auth==2.41.1

# ML inference (the deployed model is a scikit-learn RandomForest)
scikit-learn==1.7.2
joblib==1.5.2

# Data handling
pandas==2.3.3
numpy==2.2.6

# Config loading
PyYAML==6.0.3

# Charts (used in About page tech breakdown)
plotly==6.3.1
